refactor: all class files.js reformated to Capital - Pascal case class updated to pascal case
