+++ README.md
@@ -1,13 +1,14 @@
-# Sprint7 : part 5/5
+# Sprint8 : part 4/4
 
 This project includes:
-- cards.js implementation 
-- sprint 7 sub
-- Form validation
-- esc key bindings for javascript
+refactor of class js files 
+refactor of projects 
+webpack add 
+babel add 
 
-# Project 7 updated: Around The U.S.
-- sent for review #4
+
+# Project 8 updated: Around The U.S.
+- review #3
 
 ### Overview
 
@@ -44,6 +45,13 @@ To view this iteration of this first stage of the project: click here: https://z
 
 **Updates and Development**
 
+## sprint 7 updates
+This project includes:
+- cards.js implementation
+- sprint 7 sub
+- Form validation
+- esc key bindings for javascript
+
 this is revision for sprint 7
 ## Updating the project's file structure
 - Since we'll be refactoring our project using OOP, we'll update the directory structure a bit. Create two new folders: /components and /utils. The first will store your class components. The latter will store some utility functions. We'll describe their contents later.
@@ -52,8 +60,8 @@ this is revision for sprint 7
 - This is what your file structure should look like after the changes:
 - se_project_aroundtheus/
   -- components/
-      -- card.js
-      -- formValidator.js
+      -- Card.js
+      -- FormValidator.js
   -- pages/
      -- index.js
      -- index.css
@@ -76,8 +84,8 @@ this is revision for sprint 7
 This project will be different from your previous projects. Before, you were required to code and implement various features yourself, for example, adding functioning buttons or making a popup appear and then disappear. But this time, all the functionality is already there, and your task is to organize the code.
 The first requirement is to add the Card and FormValidator classes to the code and ensure that each of them performs one specific task. These classes should contain all the necessary code for performing their respective tasks.
 The second requirement is to divide the JavaScript code into modules. Your project will include four scripts:
-card.js will contain the code of the Card class.
-formValidator.js will contain the code of the FormValidator class.
+Card.js will contain the code of the Card class.
+FormValidator.js will contain the code of the FormValidator class.
 utils.js will contain the event handlers and the function that opens/closes modal windows.
 index.js will contain the rest of the code (you may have noticed that in some lessons we referred to this file as script.js; both index.js and script.js are common naming conventions for the main script file).
 The Card and FormValidator classes should be exported from their corresponding files, and then you will need to import and implement them inside of index.js.
