+++ src/pages/index.js
@@ -1,8 +1,8 @@
-import Card from '/src/components/card.js';
-import FormValidator from '/src/components/formValidator.js';
+import Card from '/src/components/Card.js';
+import FormValidator from '/src/components/FormValidator.js';
 import { openModal, closeModal, handlePopupClose } from "/src/utils/utils.js";
-import Section from '/src/utils/section.js';
-import UserInfo from '/src/utils/userInfo.js';
+import Section from '/src/utils/Section.js';
+import UserInfo from '/src/utils/UserInfo.js';
 import '/src/pages/index.css';
 import {
   initializeCards,
