+++ src/utils/PopupWithForm.js
@@ -1,4 +1,4 @@
-import Popup from './popup.js';
+import Popup from './Popup.js';
 
 class PopupWithForm extends Popup {
   constructor(popupSelector, handleFormSubmit) {
