+++ src/utils/UserInfo.js
@@ -1,4 +1,4 @@
-class userInfo {
+class UserInfo {
   constructor({ nameSelector, jobSelector }) {
     this._nameElement = document.querySelector(nameSelector);
     this._jobElement = document.querySelector(jobSelector);
